Updated TZ btn position and updated save btn.
