+++ packages/components/src/Button/IconButton/style.scss
@@ -72,6 +72,10 @@
 			border-color: transparent;
 		}
 
+		&--no-margin {
+			margin: 0;
+		}
+
 		&-color {
 			&--white {
 				background: rgba(255, 255, 255, 0.1);
