+++ packages/components/src/TimezoneTimeInfo/style.scss
@@ -1,10 +1,16 @@
 .ee-timezone-info {
 	display: inline-block;
-	margin-right: var(--ee-margin-smaller);
+	margin-left: var(--ee-margin-smaller);
 	position: relative;
 
 	&__button.ee-btn-base {
 		border: none;
+
+		svg {
+			height: 100% !important;
+			padding: var(--ee-padding-tiny);
+			width: 100% !important;
+		}
 	}
 
 	[role='dialog'] {
@@ -15,6 +21,7 @@
 		font-weight: normal;
 		left: 0;
 		min-width: 260px;
+		padding: var(--ee-padding-tiny);
 		position: absolute;
 
 		[x-arrow] {
