+++ packages/dates/src/DateRangePicker/DateRangePickerLegend.tsx
