+++ packages/dates/src/DateTimeRangePicker.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { DateRangePickerProps } from './types';
 import { DateRangePicker } from './DateRangePicker';
+import type { DateRangePickerProps } from './types';
 
 export const DateTimeRangePicker: React.FC<Omit<DateRangePickerProps, 'showTime'>> = (props) => {
 	return <DateRangePicker showTime {...props} />;
