+++ packages/dates/src/styles.scss
@@ -1,70 +1,5 @@
 @import '~@eventespresso/styles/src';
 
-:root {
-	// color variables
-	--ee-date-picker-start-date-bg-color: #a0d6b3; // #84c4bd
-	--ee-date-picker-start-date-bg-color-hover: #b5e4c6; // #ade0da
-	--ee-date-picker-date-in-range-bg-color: #9dcdd5;
-	--ee-date-picker-date-in-range-bg-color-hover: #b6d8dd;
-	--ee-date-picker-end-date-bg-color: #94b0d6;
-	--ee-date-picker-end-date-bg-color-hover: #b1c8e8;
-}
-
-.date-range-picker {
-	display: flex;
-	align-items: center;
-	&__label {
-		color: var(--ee-default-text-color-low-contrast);
-		display: block;
-		font-weight: 600;
-		margin: var(--ee-margin-small) 0 0;
-	}
-
-	&-legend {
-		&__wrapper {
-			box-sizing: border-box;
-			clear: both;
-			padding: var(--ee-padding-smaller);
-			width: 100%;
-
-			ul {
-				align-items: center;
-				display: flex;
-				flex-flow: row;
-				justify-content: space-between;
-			}
-		}
-
-		&__item {
-			align-items: center;
-			display: flex;
-			flex-flow: row;
-			justify-content: flex-start;
-			margin: 0 0 var(--ee-margin-tiny);
-
-			span {
-				height: var(--ee-icon-button-size-micro);
-				margin: 0 var(--ee-margin-tiny);
-				width: var(--ee-icon-button-size-micro);
-
-				&.legend-item {
-					&__range-start {
-						background: var(--ee-date-picker-start-date-bg-color);
-					}
-
-					&__range-day {
-						background: var(--ee-date-picker-date-in-range-bg-color);
-					}
-
-					&__range-end {
-						background: var(--ee-date-picker-end-date-bg-color);
-					}
-				}
-			}
-		}
-	}
-}
-
 /*
 	the following creates styles like:
 		.react-datepicker-popper[data-placement^="bottom"] .react-datepicker__triangle,
