+++ packages/dates/src/types.ts
@@ -13,11 +13,13 @@ export interface DatePickerProps extends Omit<ReactDatePickerProps, OmittedProps
 export type DateRange = [Date, Date];
 
 export interface DateRangePickerProps extends Omit<ReactDatePickerProps, OmittedProps> {
+	endDateTZ?: React.ReactNode;
 	endLabel?: string;
 	inputValue?: [string, string];
 	locale?: string; // "en-US", "en_US", "ar" etc.
 	onChange: (dates: DateRange) => void;
 	showTime?: boolean;
 	startLabel?: string;
+	startDateTZ?: React.ReactNode;
 	value?: DateRange;
 }
